Refactoring 03-02-2023 (#2)
* Changed the type of the id field of the Employee class from Integer to Long.

* Added support for performing unit tests agains the H2 in-memory db.

+++ src/main/java/com/example/companydbapi/controller/EmployeeController.java
@@ -25,7 +25,7 @@ public class EmployeeController {
 	}
 	
 	@GetMapping("/employees/{id}")
-	public Optional<Employee> getEmployeeById(@PathVariable(value = "id") Integer id) {
+	public Optional<Employee> getEmployeeById(@PathVariable(value = "id") Long id) {
 		return employeeRepository.findById(id);
 	}
 }
